Enable integration tests by changing to a library
Rust's integration tests work only with libraries, not with binaries.
When we generated the project, we picked the binary template, since it
best represents that Venja is an application.

Not being able to use integration tests is a serious problem, though.
The web server should be tested thoroughly, and integration tests
provide the best opportunity for that. Therefore, the decision has been
made to make the project a library, and include a binary that runs the
application.