Remove Actix
After struggling for a day to integrate Actix with Diesel and Juniper, I
want to evaluate Gotham as a replacement. Two main reasons drive this
decision:

1. Actix is currently being updated from v0.7 to v1.0. The update
   removes the dependency on the Actix framework, fundamentally changing
   the architecture. Since it is clear that v0.7 is dead, but v1.0 is
   not documented yet, building something on Actix is difficult at the
   moment.
2. Actix is quite challenging to use. It has a lot of moving parts, and
   adds another challenge. Building a web app with Rust is challenging
   enough in itself, however. Gotham has a much simpler interface, and
   can hopefully be hooked up with Diesel and Juniper much easier.

The first step to replacing Actix has been done by removing it from the
project. The next step is to add Gotham and reconfigure the database and
GraphQL.