Create middleware for Diesel
Gotham has had unofficial support for Diesel for a while. A half
finished middleware exists in its Git repository, and a work-in-progress
pull request for an example using this middleware was closed recently.

Official support seems to be a bit further down the road, as different
ways to enable proper async support are currently being discussed (see
https://github.com/gotham-rs/gotham/issues/309).

To get unblocked until official support arrives, the existing middleware
implementation has been copied into the project. The goal is to remove
it asap, but this is blocked by the above issue in Gotham.