Switch to CircleCI
Azure Pipelines was originally chosen as the CI tool, because it allowed
us to run tests without configuring a Docker based build. This made it
much easier to get started. Now that we have to run Postgres for some of
the tests, we are more or less forced to adopt Docker.

Configuring Azure is a bit of a pain, and I've gone through many
iterations trying to get the build to work. Since I'm out of ideas on
the current build failure, I want to switch to CircleCI to see if:

   a. The build error is reproducible there
   b. The configuration is easier to maintain

Since I've been using CircleCI before, it was pretty straightforward to
create a configuration. But Circle's documentation is as aweful as
Azure's...