Enable pgcrypto extension
Resources in the application have both an internal and external id. The
internal id is used for primary and foreign keys, and uses the default
integer format. This makes it performant, and easy to communicate as a
developer. But auto-incrementing integer ids leak information about the
usage of the system. That's why in all public APIs, UUIDs are used.