Wrap blocking database queries in async thread pool
Diesel implements a synchronous, blocking interface for Postgres.
Running queries in the main thread locks the tokio engine, thus
defeating the whole purpose of the async nature of Gotham.

While not officially part of Gotham yet, some people are working on an
asynchronous middleware for Diesel. The work-in-progress code has been
copied into the project, and used in the existing health check endpoint.
The original source can be found here:
https://github.com/colinbankier/gotham/tree/diesel-tokio-blocking-middleware

The new middleware uses a thread pool for synchronous calls to ensure
that the main thread doesn't get blocked.