Improve tests
The tests for the health handler have been improved. First, tests are
now run using transactions. This ensures that tests that are run in
parallel effect each other. Second, the comparison of the response body
has been refactored to make it easier to manually construct a string.
This might come in handy in the future.