Implement GraphQL endpoint
The previously empty endpoint for GraphQL has been implemented. The
implementation has been largely inspired by Juniper's existing
integrations with other web frameworks.

At this point, no useful error message is returned when an invalid
request is sent to the server. And GET requests are not support yet as
well.