Disable measuring test coverage
kcov requires certain security priviledges that are no longer available
in the build environment on CircleCI. Solutions for this problem are
based on falling back to a Unix based build environment, but this
requires more configuration. The value of test coverage is not that high
at this moment in time, so we postpone this work until a later date.