Archive project
The decision has been made to built Venja with Rails. I only have
limited time available, and it is more important to me to get this tool
built than to learn React at this point in time. The repository will be
archived and renamed to stay around for documentation and reference.